learn05/llm/requirements.txt
# 智能教学助手LLM系统依赖包

# 核心依赖
requests>=2.28.0
aiohttp>=3.8.0
pydantic>=1.10.0

# 大模型API客户端
openai>=1.0.0
dashscope>=1.14.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 系统监控
psutil>=5.9.0

# 异步支持
aiofiles>=22.1.0

# 缓存支持
redis>=4.5.0

# 日志
loguru>=0.6.0

# 配置管理
python-dotenv>=0.19.0

# 类型检查
mypy>=1.0.0

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 代码格式化
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# 开发工具
pre-commit>=2.20.0

# 可选依赖（根据需要安装）
# transformers>=4.20.0  # 本地模型支持
# torch>=1.12.0         # PyTorch支持
# tensorflow>=2.9.0     # TensorFlow支持
# scikit-learn>=1.1.0   # 机器学习支持
# matplotlib>=3.5.0     # 图表绘制
# seaborn>=0.11.0       # 统计图表
# jupyter>=1.0.0        # Jupyter支持
# fastapi>=0.95.0       # Web API框架
# uvicorn>=0.18.0       # ASGI服务器
# sqlalchemy>=1.4.0     # 数据库ORM
# alembic>=1.8.0        # 数据库迁移